feat: 리플 관련 메소드 및 API 패키지 구조 개 & 테스트코드 추가
