+++ src/main/kotlin/com/draw/controller/ReplyController.kt
@@ -7,15 +7,16 @@ import com.draw.common.Const.REPLY_TAG
 import com.draw.common.enums.Gender
 import com.draw.common.enums.MBTI
 import com.draw.controller.dto.MyRepliesRes
+import com.draw.controller.dto.ReplyCreateReq
 import com.draw.controller.dto.ReplyWriterRes
 import com.draw.service.ReplyService
 import io.swagger.v3.oas.annotations.Operation
 import io.swagger.v3.oas.annotations.Parameter
 import io.swagger.v3.oas.annotations.tags.Tag
-import mu.KotlinLogging
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.PathVariable
 import org.springframework.web.bind.annotation.PostMapping
+import org.springframework.web.bind.annotation.RequestBody
 import org.springframework.web.bind.annotation.RequestHeader
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RequestParam
@@ -27,7 +28,16 @@ import org.springframework.web.bind.annotation.RestController
 class ReplyController(
     private val replyService: ReplyService,
 ) {
-    private val log = KotlinLogging.logger { }
+
+    @PostMapping("/{feedId}/replies")
+    @Operation(summary = "리플 작성")
+    fun createReply(
+        @RequestHeader(MOCK_USER_HEADER) userId: Long,
+        @PathVariable("feedId") feedId: Long,
+        @RequestBody replyCreateReq: ReplyCreateReq,
+    ) {
+        replyService.createReply(userId, feedId, replyCreateReq)
+    }
 
     @GetMapping("/replies/me")
     @Tag(name = Const.MY_TAG, description = "My 관련 API")
