+++ src/test/kotlin/com/draw/service/ReplyServiceTest.kt
@@ -1,0 +1,35 @@
+package com.draw.service
+
+import com.draw.controller.dto.ReplyCreateReq
+import com.draw.domain.feed.Feed
+import com.draw.infra.persistence.FeedRepository
+import com.draw.infra.persistence.ReplyRepository
+import io.mockk.every
+import io.mockk.junit5.MockKExtension
+import io.mockk.mockk
+import org.assertj.core.api.Assertions.assertThat
+import org.junit.jupiter.api.Test
+import org.junit.jupiter.api.extension.ExtendWith
+import org.springframework.data.repository.findByIdOrNull
+
+@ExtendWith(value = [MockKExtension::class])
+class ReplyServiceTest {
+
+    private val feedRepository = mockk<FeedRepository>()
+    private val replyRepository = mockk<ReplyRepository>()
+
+    private val replyService = ReplyService(feedRepository, replyRepository)
+
+    @Test
+    fun `피드내 리플이 생성된다`() {
+        // given
+        val feed = Feed(content = "content", writerId = 1L)
+        every { feedRepository.findByIdOrNull(1L) } returns feed
+
+        // when
+        replyService.createReply(1L, 1L, ReplyCreateReq(content = "content"))
+
+        // then
+        assertThat(feed.replies).hasSize(1)
+    }
+}
